extracting function out of MultipleChoice
